Move auth controller logic into the routes
Done to be able to harness type inference more easily, and to couple
the logic relating to validation and typing more closely.
